+++ src/main/java/com/balsamic/sejongmalsami/controller/DocumentPostController.java
@@ -4,7 +4,6 @@ import com.balsamic.sejongmalsami.object.CustomUserDetails;
 import com.balsamic.sejongmalsami.object.DocumentCommand;
 import com.balsamic.sejongmalsami.object.DocumentDto;
 import com.balsamic.sejongmalsami.service.DocumentPostService;
-import com.balsamic.sejongmalsami.service.PopularPostService;
 import com.balsamic.sejongmalsami.util.log.LogMonitoringInvocation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
@@ -26,8 +25,6 @@ import org.springframework.web.bind.annotation.RestController;
 public class DocumentPostController implements DocumentPostControllerDocs {
 
   private final DocumentPostService documentPostService;
-  private final PopularPostService popularPostService;
-
 
   @Override
   @PostMapping(value = "/post", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
@@ -37,20 +34,4 @@ public class DocumentPostController implements DocumentPostControllerDocs {
       @ModelAttribute DocumentCommand command) {
     return ResponseEntity.ok(documentPostService.saveDocumentPost(command));
   }
-
-  @Override
-  @PostMapping(value = "/daily/popular", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-  @LogMonitoringInvocation
-  public ResponseEntity<DocumentDto> getDailyPopularDocumentPost(
-      @ModelAttribute DocumentCommand command) {
-    return ResponseEntity.ok(popularPostService.getDailyPopularDocumentPosts());
-  }
-
-  @Override
-  @PostMapping(value = "/weekly/popular", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-  @LogMonitoringInvocation
-  public ResponseEntity<DocumentDto> getWeeklyPopularDocumentPost(DocumentCommand command) {
-    return ResponseEntity.ok(popularPostService.getWeeklyPopularDocumentPosts());
-  }
-
 }
