+++ src/main/java/com/balsamic/sejongmalsami/controller/DocumentPostControllerDocs.java
@@ -80,80 +80,4 @@ public interface DocumentPostControllerDocs {
   ResponseEntity<DocumentDto> saveDocumentPost(
       CustomUserDetails customUserDetails,
       DocumentCommand command);
-
-  @ApiChangeLogs({
-      @ApiChangeLog(
-          date = "2024.10.15",
-          author = Author.BAEKJIHOON,
-          description = "dto 필드 변경에 따른 반환값 수정"
-      ),
-      @ApiChangeLog(
-          date = "2024.10.10",
-          author = Author.BAEKJIHOON,
-          description = "자료게시판 일간 인기글 init"
-      )
-  })
-  @Operation(
-      summary = "일간 인기 자료글",
-      description = """
-          **자료 일간 인기글 요청**
-
-          **이 API는 인증이 필요하며, JWT 토큰이 존재해야합니다.**
-
-          **입력 파라미터 값:**
-
-          없음
-
-          **반환 파라미터 값:**
-
-          - **DocumentDto**: 자료 게시판 정보 반환
-            - **List\\<DocumentPost\\> documentPosts**: 일간 자료 인기글 리스트
-
-          **참고 사항:**
-
-          - 이 API를 통해 사용자는 일간 인기 자료글을 조회할 수 있습니다.
-          - 요청 시각으로부터 24시간 이내에 작성된 상위 30개의 일간 인기글을 조회합니다.
-          - Swagger에서 테스트 시 mediaFiles에 있는 "Send empty value" 체크박스 해제해야합니다.
-          """
-  )
-  ResponseEntity<DocumentDto> getDailyPopularDocumentPost(
-      DocumentCommand command);
-
-  @ApiChangeLogs({
-      @ApiChangeLog(
-          date = "2024.10.15",
-          author = Author.BAEKJIHOON,
-          description = "dto 필드 변경에 따른 반환값 수정"
-      ),
-      @ApiChangeLog(
-          date = "2024.10.10",
-          author = Author.BAEKJIHOON,
-          description = "자료게시판 주간 인기글 init"
-      )
-  })
-  @Operation(
-      summary = "주간 인기 자료글",
-      description = """
-          **자료 주간 인기글 요청**
-
-          **이 API는 인증이 필요하며, JWT 토큰이 존재해야합니다.**
-
-          **입력 파라미터 값:**
-
-          없음
-
-          **반환 파라미터 값:**
-
-          - **DocumentDto**: 자료 게시판 정보 반환
-            - **List\\<DocumentPost\\> documentPosts**: 주간 자료 인기글 리스트
-
-          **참고 사항:**
-
-          - 이 API를 통해 사용자는 주간 인기 자료글을 조회할 수 있습니다.
-          - 요청 시각으로부터 7일 이내에 작성된 상위 30개의 주간 인기글을 조회합니다.
-          - Swagger에서 테스트 시 mediaFiles에 있는 "Send empty value" 체크박스 해제해야합니다.
-          """
-  )
-  ResponseEntity<DocumentDto> getWeeklyPopularDocumentPost(
-      DocumentCommand command);
 }
